refactor(svg): extract tower layout + animation helpers
