+++ lib/svg/generator.ts
@@ -1,67 +1,13 @@
 import type { BadgeParams, ContributionCalendar, StreakStats } from '../../types';
 import { AUTO_DARK_THEME, AUTO_LIGHT_THEME } from './themes';
-
-// constants
-const GHOST_HEIGHT_PX = 4;
-const LOG_SCALE_MULTIPLIER = 12;
-const LINEAR_SCALE_MULTIPLIER = 5;
-const MAX_LOG_HEIGHT = 80;
-const MAX_LINEAR_HEIGHT = 50;
-
-// TOWER_BASE_Y: the vertical midpoint of the isometric ground floor diamond in local SVG space.
-// The diamond paths are drawn from y=0 (back vertex) to y=20 (front vertex), making y=10
-// the horizontal center line that acts as the visual ground level for each tower.
-// This value is used as the CSS transform-origin for the grow-up animation so towers
-// scale upward from their ground tile rather than from the SVG origin.
-const TOWER_BASE_Y = 10;
-
-// Shared animation CSS injected into both static and auto-theme SVG renderers.
-// Defined once here to avoid duplication — any curve/timing change only needs updating in one place.
-const TOWER_ANIMATION_CSS = `
-  .cp-tower {
-    transform: scaleY(0);
-    transform-origin: 0 ${TOWER_BASE_Y}px;
-    animation: grow-up 1.2s cubic-bezier(0.16, 1, 0.3, 1) forwards;
-  }
-  @keyframes grow-up {
-    from { transform: scaleY(0); }
-    to   { transform: scaleY(1); }
-  }
-  @media (prefers-reduced-motion: reduce) {
-    .cp-tower { animation: none !important; transform: scaleY(1) !important; }
-  }`;
+import { TOWER_ANIMATION_CSS } from './animations';
+import { computeTowers } from './layout';
 
 const FONT_MAP: Record<string, string> = {
   jetbrains: '"JetBrains Mono", monospace',
   fira: '"Fira Code", monospace',
   roboto: '"Roboto", sans-serif',
 };
-
-// types
-/** Shared layout data for a single isometric tower. */
-interface FaceOpacity {
-  left: number;
-  right: number;
-  top: number;
-}
-
-interface TowerData {
-  x: number;
-  y: number;
-  h: number;
-  hasCommits: boolean;
-  isGhost: boolean;
-  isToday: boolean;
-  isTodayWithCommits: boolean;
-  tooltip: string;
-  contributionCount: number;
-  faceOpacity: FaceOpacity;
-  strokeOpacity: number;
-  strokeWidth: number;
-  /** Grid position used to compute the staggered animation-delay (row + col) * offset */
-  row: number;
-  col: number;
-}
 
 // helpers
 function deterministicRandom(seed: string): number {
@@ -71,95 +17,6 @@ function deterministicRandom(seed: string): number {
     hash = Math.imul(hash, 16777619);
   }
   return (hash >>> 0) / 4294967296;
-}
-function computeTowerHeight(
-  count: number,
-  scale: 'linear' | 'log',
-  shouldShowGhostCity: boolean
-): number {
-  if (count === 0 && shouldShowGhostCity) return GHOST_HEIGHT_PX;
-  if (count === 0) return 0;
-  return scale === 'log'
-    ? Math.min(Math.log2(count + 1) * LOG_SCALE_MULTIPLIER, MAX_LOG_HEIGHT)
-    : Math.min(count * LINEAR_SCALE_MULTIPLIER, MAX_LINEAR_HEIGHT);
-}
-
-function computeFaceOpacity(count: number, isGhostCityMode: boolean): FaceOpacity {
-  if (isGhostCityMode) {
-    return { left: 0, right: 0, top: 0.02 };
-  }
-  if (count === 0) {
-    return { left: 0, right: 0, top: 0.02 };
-  }
-  return { left: 0.35, right: 0.21, top: 0.7 };
-}
-
-/**
- * Computes tower positions and heights from the last 14 weeks of
- * contribution data. The layout math is identical for both the
- * static-theme and auto-theme rendering paths.
- */
-function computeTowers(
-  calendar: ContributionCalendar,
-  scale: 'linear' | 'log',
-  todayDate: string
-): TowerData[] {
-  const weeks = calendar.weeks.slice(-14);
-  const towers: TowerData[] = [];
-
-  // Calculate if the entire monolith is empty
-  let totalVisibleContributions = 0;
-  weeks.forEach((week) => {
-    week.contributionDays.forEach((day) => {
-      totalVisibleContributions += day.contributionCount;
-    });
-  });
-
-  const shouldShowGhostCity = totalVisibleContributions === 0;
-
-  // Pre-check: is todayDate present in the visible 14-week window?
-  // If not (e.g. stale cache or todayDate outside the window), fall back to
-  // marking the last visible day as "today" so the pulse always appears.
-  const todayInWindow = weeks.some((w) => w.contributionDays.some((d) => d.date === todayDate));
-
-  weeks.forEach((week, i) => {
-    week.contributionDays.forEach((day, j) => {
-      // Use the caller-supplied local date so the pulse animation fires on the
-      // correct tower for users in non-UTC timezones, not always the last UTC entry.
-      const isToday =
-        day.date === todayDate ||
-        // Fallback: if todayDate isn't in the visible window, keep the old behaviour.
-        (!todayInWindow && i === weeks.length - 1 && j === week.contributionDays.length - 1);
-      const hasCommits = day.contributionCount > 0;
-      const isGhost = !hasCommits && shouldShowGhostCity;
-      const isTodayWithCommits = isToday && hasCommits;
-
-      const tooltip = isTodayWithCommits
-        ? `TODAY: ${day.date}: ${day.contributionCount} contributions`
-        : `${day.date}: ${day.contributionCount} contributions`;
-
-      // If not ghost city and no commits, height is 0, so don't render face if not needed,
-      // but we return 0 for height so it won't be visible.
-      towers.push({
-        x: 300 + (i - j) * 16,
-        y: 120 + (i + j) * 9,
-        h: computeTowerHeight(day.contributionCount, scale, shouldShowGhostCity),
-        hasCommits,
-        isGhost,
-        isToday,
-        isTodayWithCommits,
-        tooltip,
-        contributionCount: day.contributionCount,
-        faceOpacity: computeFaceOpacity(day.contributionCount, shouldShowGhostCity),
-        strokeOpacity: isGhost ? 0.3 : 0,
-        strokeWidth: isGhost ? 0.5 : 0,
-        row: i,
-        col: j,
-      });
-    });
-  });
-
-  return towers;
 }
 
 function escapeXML(str: string): string {
